requirements.txt
google-genai
python-dotenv
sqlalchemy
PyYAML
fastapi
uvicorn[standard]
pydantic
chromadb